pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Dmitry_Perexozhev <perexozhev99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
validators = "^0.33.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.script:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
flake8 = "^7.1.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
